% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_colSums}
\alias{par_colSums}
\title{Calculate the column sums for each column of a matrix.
OpenMP is used to calculate improve calculation times.
TODO: Needs to implement a multicore cliff.}
\usage{
par_colSums(x, na_rm = FALSE, threads = 1L, display_progress = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{na_rm}{Does nothing at present.}

\item{threads}{The number of threads to run calculation over}
}
\description{
Calculate the column sums for each column of a matrix.
OpenMP is used to calculate improve calculation times.
TODO: Needs to implement a multicore cliff.
}
