% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_colVars}
\alias{par_colVars}
\title{Calculate the column variances of a matrix
* OpenMP is used to calculate improve calculation times.
TODO: Needs to implement a multicore cliff.}
\usage{
par_colVars(x, na_rm = FALSE, threads = 1L)
}
\arguments{
\item{x}{A numeric vector}

\item{na_rm}{Does nothing at present.}

\item{threads}{The number of threads to run calculation over}
}
\description{
Calculate the column variances of a matrix
* OpenMP is used to calculate improve calculation times.
TODO: Needs to implement a multicore cliff.
}
