% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{par_cor}
\alias{par_cor}
\title{Correlation using OpenMP}
\usage{
par_cor(x, y = NULL, use = "everything", method = "pearson",
  threads = 1L)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A nueric vector}

\item{use}{Not Implemented}

\item{method}{Method for calcuating the correlation. Curreltly supports Pearson and Spearman correlations.}

\item{threads}{Threads to use for OpenMP}
}
\description{
Correlation using OpenMP
}
