% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_dist_old}
\alias{par_dist_old}
\title{Matrix Distance Measures}
\usage{
par_dist_old(x, method = "euclidian", na_rm = FALSE, threads = 1L,
  display_progress = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{na_rm}{Does nothing at present.}

\item{threads}{The number of threads to run calculation over}
}
\description{
Calculate the distance bewteen matrix rows using various distance measures.
}
\details{
OpenMP is used to calculate improve calculation times.
}
