% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_mean}
\alias{par_mean}
\title{Calculate the mean of a numeric vector
OpenMP is used to calculate improve calculation times.}
\usage{
par_mean(x, na_rm = FALSE, threads = 1L)
}
\arguments{
\item{x}{A numeric vector}

\item{na_rm}{Does nothing at present.}

\item{threads}{The number of threads to run calculation over}
}
\description{
Calculate the mean of a numeric vector
OpenMP is used to calculate improve calculation times.
}
