% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_sort}
\alias{par_sort}
\title{* Sort a vector using __gnu_parallel::sort.
* Code is not parallel, but is more efficient for numeric vectors than base R.
Parallel sort codes is https://gcc.gnu.org/onlinedocs/libstdc++/manual/parallel_mode_using.html}
\usage{
par_sort(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
* Sort a vector using __gnu_parallel::sort.
* Code is not parallel, but is more efficient for numeric vectors than base R.
Parallel sort codes is https://gcc.gnu.org/onlinedocs/libstdc++/manual/parallel_mode_using.html
}
