% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_sum}
\alias{par_sum}
\title{Vector and Matrix Sums}
\usage{
par_sum(x, na_rm = FALSE, threads = 1L)
}
\arguments{
\item{x}{A numeric vector}

\item{na_rm}{Does nothing at present.}

\item{threads}{The number of threads to run calculation over}
}
\description{
Calculate the sum of a vector
}
\details{
Calculate the sum of a numeric vector
OpenMP is used to calculate improve calculation times.
}
