% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_var}
\alias{par_var}
\title{Calculate the varance of a vector.
Currently uses the Rcpp sugar implementation of the variance function.
TODO: Needs to implement a multicore cliff.
Based on https://en.wikipedia.org/wiki/Algorithms_for_calculating_variance}
\usage{
par_var(x, na_rm = FALSE, threads = 1L)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Calculate the varance of a vector.
Currently uses the Rcpp sugar implementation of the variance function.
TODO: Needs to implement a multicore cliff.
Based on https://en.wikipedia.org/wiki/Algorithms_for_calculating_variance
}
